# Adding a ShuffleInPlace that takes in a Span