Just adding an overload to ShuffleInPlace for Spans and forwarding to it when possible from the IList<T> version. I saw fairly significant performance improvements when benchmarking it with 1000 ints on my system:

Before:
| Method       | Job                | Runtime            | Mean      | Error    | StdDev   |
|------------- |------------------- |------------------- |----------:|---------:|---------:|
| ShuffleList  | .NET 6.0           | .NET 6.0           | 218.76 ms | 2.931 ms | 2.448 ms |
| ShuffleArray | .NET 6.0           | .NET 6.0           | 257.76 ms | 1.906 ms | 1.783 ms |
| ShuffleList  | .NET 8.0           | .NET 8.0           |  62.22 ms | 1.119 ms | 1.046 ms |
| ShuffleArray | .NET 8.0           | .NET 8.0           | 184.37 ms | 2.613 ms | 2.182 ms |
| ShuffleList  | .NET Framework 4.8 | .NET Framework 4.8 | 260.43 ms | 5.103 ms | 5.877 ms |
| ShuffleArray | .NET Framework 4.8 | .NET Framework 4.8 | 260.39 ms | 2.187 ms | 2.046 ms |

After:
| Method       | Job                | Runtime            | Mean      | Error    | StdDev   |
|------------- |------------------- |------------------- |----------:|---------:|---------:|
| ShuffleList  | .NET 6.0           | .NET 6.0           |  93.59 ms | 1.476 ms | 1.233 ms |
| ShuffleArray | .NET 6.0           | .NET 6.0           |  89.88 ms | 1.224 ms | 1.085 ms |
| ShuffleList  | .NET 8.0           | .NET 8.0           |  52.50 ms | 0.439 ms | 0.343 ms |
| ShuffleArray | .NET 8.0           | .NET 8.0           |  53.42 ms | 1.051 ms | 1.329 ms |
| ShuffleList  | .NET Framework 4.8 | .NET Framework 4.8 | 243.63 ms | 2.525 ms | 2.109 ms |
| ShuffleArray | .NET Framework 4.8 | .NET Framework 4.8 |  88.02 ms | 1.710 ms | 2.035 ms |